@@ -1,13 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-	compileSdkVersion 26
-	buildToolsVersion "26.0.2"
+	compileSdkVersion 30
 
 	defaultConfig {
 		applicationId "com.cube.storm.util.example"
 		minSdkVersion 16
-		targetSdkVersion 26
+		targetSdkVersion 30
 		versionCode 1
 		versionName "1.0"
 	}
@@ -18,7 +17,7 @@ android {
 }
 
 dependencies {
-	compile fileTree(dir: 'libs', include: ['*.jar'])
-	compile 'com.android.support:appcompat-v7:26.0.2'
-	compile project(':library')
+	implementation fileTree(dir: 'libs', include: ['*.jar'])
+	implementation 'androidx.appcompat:appcompat:1.1.0'
+	implementation project(':library')
 }
