@@ -5,9 +5,9 @@
  **/
 package com.cube.storm.util.lib.debug;
 
-import android.support.annotation.IntDef;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.IntDef;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 
 import com.cube.storm.util.BuildConfig;
