@@ -2,7 +2,7 @@ package com.cube.storm.util.lib.resolver;
 
 import android.content.Context;
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import java.io.IOException;
