@@ -1,7 +1,7 @@
 package com.cube.storm.util.lib.resolver;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.io.File;
 import java.io.FileInputStream;
